requirements.txt
torchmetrics
transformers
tqdm
pillow
numpy
scikit-learn
matplotlib
pandas
kornia
ipykernel
torch-fidelity